Revert the testsuite changes from previous commit
Can't BuildRequire wine without breaking the build on arches where wine
is unavailable. As such, the testsuite should only be used for local
testing and not enabled in official builds.

Even on secondary arches, all the builds are done from a single source
RPM that gets constructed on the primary koji instance. Because of the
single SRPM issue, using build time conditionals won't work for
conditionally selecting BuildRequires -- the SRPM requires will be
chosen depending on where the SRPM happens to get built and the same set
of requires will also be used an all other arches.